Edits to mission, adds "active" to Maintainer role
Line 11: Suggested edits to the Mission to 1) broaden the advocacy/tools to general vuln disclosure and 2) set the tone around tooling focus (ie "solve gaps"). 

Line 29: Adds the idea that Maintainers are expected to be "active" in the WG to retain their Maintainer status. Active is intentionally undefined and -- in the ethos of "minimum viable governance" -- could be defined by the TSC if this becomes an issue. 

Signed-off-by: Anne Bertucio <annebertucio@google.com>